Refactor reservation logic into `ReservationService` and update controller to use the new service.
